requirements.txt
pandas==1.5.2
requests==2.28.1
psycopg2-binary==2.9.5